command-center/packages/agent-server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
anthropic>=0.40.0
pyodbc>=5.1.0
pymssql>=2.2.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
pydantic>=2.7.0
sse-starlette>=2.0.0
